refactor: Rename pipelines classes to Base and move it to base module.
