+++ filejacket/pipelines/renamer.py
@@ -24,169 +24,19 @@ Should there be a need for contact the electronic mail
 # Python internals
 from __future__ import annotations
 
-import logging
 import re
-from typing import Any
-from typing import Type, Pattern, TYPE_CHECKING
+from typing import Pattern
 from uuid import uuid4
 
 # modules
-from ..engines.storage import Storage
+from .base import BaseRenamer
 
-if TYPE_CHECKING:
-    from ..file import BaseFile
 
 __all__ = [
-    'BaseRenamer',
     'WindowsRenamer',
     'LinuxRenamer',
     'UniqueRenamer'
 ]
-
-
-class BaseRenamer:
-    """
-    Base class to be inherent to define class to be used on Renamer pipeline.
-    """
-
-    stopper: bool = True
-    """
-    Variable that define if this class used as processor should stop the pipeline.
-    """
-    file_system_handler: Type[Storage] = Storage
-    """
-    Variable to store the local storage system.
-    """
-    enumeration_pattern: Pattern
-    enumeration_pattern = None
-    """
-    Variable that define the pattern to detect enumeration in file system.
-    """
-
-    reserved_names: list = []
-    """
-    Variable that define the reversed names by the file system that should be avoid in renaming.
-    """
-
-    @classmethod
-    def prepare_filename(cls, filename: str, extension: str | None = None) -> tuple[str, str | None]:
-        """
-        Method to separated extension from filename if extension
-        not informed and save on class.
-        """
-        # Remove extension from filename if it was given
-        if extension and filename[-len(extension):] == extension:
-            filename = filename[:-len(f".{extension}")]
-
-        return filename, extension
-
-    @classmethod
-    def get_name(cls, directory_path: str, filename: str, extension: str | None) -> tuple[str, str | None]:
-        """
-        Method to get the new generated name.
-        This class should raise BlockingIOError when a custom error should happen that will be
-        caught by `process` when using Pipeline.
-        """
-        raise NotImplementedError("Method get_name must be overwrite on child class.")
-
-    @classmethod
-    def process(cls, **kwargs: Any) -> bool:
-        """
-        Method used to run this class on Processor`s Pipeline for Files.
-        This method and to_processor() is not need to rename files outside a pipeline.
-        This process method is created exclusively to pipeline for objects inherent from BaseFile.
-
-        The processor for renamer uses only one object that must be settled through first argument
-        or through key work `object`.
-
-        The keyword argument `path_attribute` allow using a different attribute for specify the file object path other
-        than the default `path`.
-        The keyword argument `reserved_names` allow for override of current list of reserved_names in pipeline. This
-        override will affect the class and thus all usage of `reserved_names`. It isn`t thread safe.
-
-        FUTURE CONSIDERATION: Making the pipeline multi thread or multi process only will required that
-        a lock be put between usage of get_name.
-        FUTURE CONSIDERATION: Multi thread will need to consider that the attribute `file_system_handler`
-        is shared between the reference of the class and all object of it and will have to be change the
-        code (multi process don't have this problem).
-
-        This processors return boolean to indicate that process was ran successfully.
-
-        This method can throw BlockingIOError when trying to rename the file.
-        The `Pipeline.run` method will catch it.
-        """
-        # Get default values from keywords arguments
-        object_to_process: BaseFile = kwargs.pop('object_to_process')
-        path_attribute: str = kwargs.pop('path_attribute', 'path')
-        reserved_names: list | None = kwargs.pop('reserved_names', None)
-
-        # Override current reserved names if list of new one provided.
-        if reserved_names:
-            cls.clean_reserved_names()
-            cls.add_reserved_name(reserved_names)
-
-        # Prepare filename from File's object
-        filename, extension = cls.prepare_filename(object_to_process.filename, object_to_process.extension)
-
-        # Save current file system filejacket
-        class_file_system_handler = cls.file_system_handler
-
-        # Overwrite File System attribute with File System of File only when running in pipeline.
-        # This will alter the File System for the class, any other call to this class will use the altered
-        # file system.
-        cls.file_system_handler = object_to_process.storage
-
-        # Get directory from object to be processed.
-        path = cls.file_system_handler.sanitize_path(getattr(object_to_process, path_attribute))
-
-        # When is not possible to get new name by some problem either with file or filesystem
-        # is expected BlockingIOError
-        new_filename, extension = cls.get_name(path, filename, extension)
-
-        # Restore File System attribute to original.
-        cls.file_system_handler = class_file_system_handler
-
-        # Set new name at File's object.
-        # The File class should set the old name at File`s cache/history automatically,
-        # filename and extension should be property functions.
-        object_to_process.complete_filename_as_tuple = (new_filename, extension)
-
-        return True
-
-    @classmethod
-    def is_name_reserved(cls, filename: str, extension: str) -> bool:
-        """
-        Method to check if filename in list of reserved names.
-        Those name should be set-up before rename pipeline being called.
-        """
-        return filename + extension in cls.reserved_names
-
-    @classmethod
-    def add_reserved_name(cls, value: str | list) -> None:
-        """
-        Method to update list of reserved names allowing append of multiple values with list.
-        This method accept string or list to be added to reserved_names.
-        """
-        if isinstance(value, str):
-            cls.reserved_names.append(value)
-
-        elif isinstance(value, list):
-            cls.reserved_names += value
-
-    @classmethod
-    def clean_reserved_names(cls) -> None:
-        """
-        Method to reset the `reserved_names` attribute to a empty list.
-        """
-        cls.reserved_names = []
-
-    @classmethod
-    def register_error(cls, error: Exception) -> None:
-        """
-        Method to log error in system. It could be override to register error in list or distinct output.
-        """
-        # Log error message
-        logging.error(str(error))
 
 
 class WindowsRenamer(BaseRenamer):
